requirements.txt
aniso8601==6.0.0
chardet==3.0.4
Click==7.0
codecov==2.0.15
coverage==4.5.3
dnspython==1.16.0
eventlet==0.24.1
Flask==1.0.2
Flask-Cors==3.0.7
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
gitdb2==2.0.5
GitPython==2.1.11
gnenv==0.1.3
greenlet==0.4.15
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
monotonic==1.5
nose==1.3.7
nose-parameterized==0.6.0
numpy==1.16.2
Pillow==6.0.0
psycopg2-binary==2.8.2
python-dateutil==2.8.0
python-slugify==1.2.6
pytz==2019.1
PyYAML==5.1
requests==2.21.0
scipy==1.2.1
six==1.12.0
smmap2==2.0.5
SQLAlchemy==1.3.2
statsd==3.3.0
strict-rfc3339==0.7
typing==3.6.6
Unidecode==1.0.22
urllib3==1.24.1
Werkzeug==0.15.2
Yapsy==1.12.0
